# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Excelsior Deckbuilder API v1
  description: |
    External HTTP API for the Excelsior Overpower card database and deck builder.

    Stability: this is a stub spec. Paths and payloads are authoritative from
    [`API_V1.md`](../API_V1.md); this document is maintained alongside it so
    external clients can generate SDKs. Any new `/api/v1` route must add an
    entry here, an error-catalog entry, and a changelog entry in the same PR.

    Contract highlights:

      - Responses use the v1 envelope `{ data, meta, errors }`.
      - Authentication is `Bearer <accessToken>` JWT (see `/auth/login`).
      - Refresh tokens rotate on `/auth/refresh`; reuse revokes the family.
      - Rate limits are per-user (or per-IP for unauthenticated calls). See
        `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`.
      - Catalog endpoints are edge-cached with `Cache-Control` + `ETag`; clients
        SHOULD send `If-None-Match` to get `304 Not Modified`.
      - Images are served from `APP_CDN_BASE` over HTTPS. See
        [`API_V1_IMAGE_CONTRACT.md`](current/API_V1_IMAGE_CONTRACT.md).
  version: 1.0.0
  contact:
    name: Excelsior
    url: https://excelsior.cards
servers:
  - url: https://excelsior.cards
    description: Production
  - url: http://localhost:8085
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DeckPreviewDraft:
      type: object
      required:
        - schemaVersion
        - draftId
        - revision
        - cards
      additionalProperties: false
      properties:
        schemaVersion:
          type: integer
          enum:
            - 1
        draftId:
          type: string
          minLength: 1
          maxLength: 100
        revision:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        cards:
          type: array
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required:
              - type
              - cardId
              - quantity
            properties:
              type:
                type: string
                enum:
                  - character
                  - special
                  - power
                  - mission
                  - event
                  - location
                  - battleground
                  - aspect
                  - advanced-universe
                  - advanced_universe
                  - teamwork
                  - ally-universe
                  - ally_universe
                  - training
                  - basic-universe
                  - basic_universe
              cardId:
                type: string
                minLength: 1
                maxLength: 100
              quantity:
                type: integer
                minimum: 1
                maximum: 100
              exclude_from_draw:
                type: boolean
        reserveCharacterId:
          type: string
          nullable: true
        limited:
          type: boolean
          default: false
        format:
          type: string
          enum:
            - venture
          default: venture
        koCharacterIds:
          type: array
          maxItems: 100
          items:
            type: string
    DrawDraftRequest:
      type: object
      additionalProperties: false
      required:
        - draft
      properties:
        draft:
          $ref: '#/components/schemas/DeckPreviewDraft'
    ExportDraftRequest:
      type: object
      additionalProperties: false
      required:
        - draft
        - display
      properties:
        draft:
          $ref: '#/components/schemas/DeckPreviewDraft'
        display:
          type: object
          additionalProperties: false
          required:
            - name
            - description
            - exportedBy
            - surface
          properties:
            name:
              type: string
              maxLength: 100
            description:
              type: string
              maxLength: 500
            exportedBy:
              type: string
              maxLength: 100
            surface:
              type: string
              enum:
                - editor
                - selection
    AnalyzeHandRequest:
      type: object
      additionalProperties: false
      required:
        - draft
        - hand
      properties:
        draft:
          $ref: '#/components/schemas/DeckPreviewDraft'
        hand:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - type
              - cardId
            properties:
              type:
                type: string
                minLength: 1
                maxLength: 40
              cardId:
                type: string
                minLength: 1
                maxLength: 100
                pattern: ^[a-zA-Z0-9_-]+$
          maxItems: 9
    DeckSummariesRequest:
      type: object
      additionalProperties: false
      required:
        - drafts
      properties:
        drafts:
          type: array
          items:
            $ref: '#/components/schemas/DeckPreviewDraft'
          minItems: 1
          maxItems: 20
      description: At most 500 typed card identities across this batch; excess is VALIDATION_ERROR.
...
    DrawDraftResult:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - inputKey
        - revision
        - cards
      properties:
        schemaVersion:
          type: integer
          enum:
            - 1
        inputKey:
          type: string
        revision:
          type: integer
          minimum: 0
        cards:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - type
              - cardId
              - quantity
            properties:
              type:
                type: string
                enum:
                  - character
                  - special
                  - power
                  - mission
                  - event
                  - location
                  - battleground
                  - aspect
                  - advanced-universe
                  - teamwork
                  - ally-universe
                  - training
                  - basic-universe
              cardId:
                type: string
              quantity:
                type: number
              exclude_from_draw:
                type: boolean
              instanceId:
                type: string
          maxItems: 9
    HandAnalysisResult:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - inputKey
        - revision
        - ventureTotal
        - duplicateCount
        - duplicateCardIndexes
      properties:
        schemaVersion:
          type: integer
          enum:
            - 1
        inputKey:
          type: string
        revision:
          type: integer
          minimum: 0
        ventureTotal:
          type: number
        duplicateCount:
          type: number
        duplicateCardIndexes:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 8
    DeckExportResult:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - inputKey
        - revision
        - deck
      properties:
        schemaVersion:
          type: integer
          enum:
            - 1
        inputKey:
          type: string
        revision:
          type: integer
          minimum: 0
        deck:
          type: object
          additionalProperties: false
          required:
            - name
            - description
            - export_timestamp
            - exported_by
            - total_cards
            - max_energy
            - max_combat
            - max_brute_force
            - max_intelligence
            - total_energy_icons
            - total_combat_icons
            - total_brute_force_icons
            - total_intelligence_icons
            - total_threat
            - reserve_character
            - cataclysm_special
            - assist_special
            - ambush_special
            - legal
            - limited
            - cards
          properties:
            name:
              type: string
            description:
              type: string
            export_timestamp:
              type: string
            exported_by:
              type: string
            total_cards:
              type: number
            max_energy:
              type: number
            max_combat:
              type: number
            max_brute_force:
              type: number
            max_intelligence:
              type: number
            total_energy_icons:
              type: number
            total_combat_icons:
              type: number
            total_brute_force_icons:
              type: number
            total_intelligence_icons:
              type: number
            total_threat:
              type: number
            reserve_character:
              type: string
              nullable: true
            cataclysm_special:
              type: string
              nullable: true
            assist_special:
              type: string
              nullable: true
            ambush_special:
              type: string
              nullable: true
            legal:
              type: boolean
            limited:
              type: boolean
            cards:
              type: object
              additionalProperties: false
              required:
                - characters
                - locations
                - battlegrounds
                - aspects
                - teamwork
                - allies
                - training
                - basic_universe
                - power_cards
                - special_cards
                - missions
                - events
                - advanced_universe
              properties:
                characters:
                  type: array
                  items:
                    type: string
                locations:
                  type: array
                  items:
                    type: string
                battlegrounds:
                  type: array
                  items:
                    type: string
                aspects:
                  type: array
                  items:
                    type: string
                teamwork:
                  type: array
                  items:
                    type: string
                allies:
                  type: array
                  items:
                    type: string
                training:
                  type: array
                  items:
                    type: string
                basic_universe:
                  type: array
                  items:
                    type: string
                power_cards:
                  type: array
                  items:
                    type: string
                special_cards:
                  type: object
                  additionalProperties:
                    type: array
                    items:
                      type: string
                missions:
                  type: object
                  additionalProperties:
                    type: array
                    items:
                      type: string
                events:
                  type: object
                  additionalProperties:
                    type: array
                    items:
                      type: string
                advanced_universe:
                  type: object
                  additionalProperties:
                    type: array
                    items:
                      type: string
    DeckSummaryResult:
      type: object
      additionalProperties: false
      required:
        - draftId
        - inputKey
        - grid
      properties:
        draftId:
          type: string
        inputKey:
          type: string
        grid:
          type: object
          additionalProperties: false
          required:
            - energy
            - combat
            - bruteForce
            - intelligence
          properties:
            energy:
              type: number
            combat:
              type: number
            bruteForce:
              type: number
            intelligence:
              type: number
          nullable: true
...
paths:
...
  /decks/draw:
    post:
      summary: Draw a fresh random hand without persisting a deck
      description: >-
        Cache-Control: no-store. Shared preview budget 120/minute. Public preview POSTs are decks:read service operations;
        administration remains unavailable to the application service adapter.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DrawDraftRequest'
      responses:
        '200':
          description: V1 envelope; no deck or collection writes. Draw/export/analysis echo exact revision and inputKey.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DrawDraftResult'
                  meta:
                    type: object
                  errors:
                    type: array
                    items:
                      type: object
        '400':
          description: VALIDATION_ERROR or DRAFT_STRUCTURE_INVALID
        '429':
          description: RATE_LIMITED
        '503':
          description: DRAFT_PREVIEW_UNAVAILABLE; retain draft, explicit retry
      security: []
  /decks/export:
    post:
      summary: Calculate the characterized export from a server-resolved draft
      description: >-
        Cache-Control: no-store. Shared preview budget 120/minute. Public preview POSTs are decks:read service operations;
        administration remains unavailable to the application service adapter.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportDraftRequest'
      responses:
        '200':
          description: V1 envelope; no deck or collection writes. Draw/export/analysis echo exact revision and inputKey.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DeckExportResult'
                  meta:
                    type: object
                  errors:
                    type: array
                    items:
                      type: object
        '400':
          description: VALIDATION_ERROR or DRAFT_STRUCTURE_INVALID
        '429':
          description: RATE_LIMITED
        '503':
          description: DRAFT_PREVIEW_UNAVAILABLE; retain draft, explicit retry
      security: []
  /decks/summaries:
    post:
      summary: Evaluate maximum grids for a bounded batch without persistence
      description: >-
        Cache-Control: no-store. Shared preview budget 120/minute. Public preview POSTs are decks:read service operations;
        administration remains unavailable to the application service adapter.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeckSummariesRequest'
      responses:
        '200':
          description: V1 envelope; no deck or collection writes. Draw/export/analysis echo exact revision and inputKey.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeckSummaryResult'
                  meta:
                    type: object
                  errors:
                    type: array
                    items:
                      type: object
        '400':
          description: VALIDATION_ERROR or DRAFT_STRUCTURE_INVALID
        '429':
          description: RATE_LIMITED
        '503':
          description: DRAFT_PREVIEW_UNAVAILABLE; retain draft, explicit retry
      security: []
  /admin/decks/hand-analysis:
    post:
      summary: Analyze available hand copies using a verified ADMIN player
      description: >-
        Cache-Control: no-store. Shared preview budget 120/minute. Public preview POSTs are decks:read service operations;
        administration remains unavailable to the application service adapter.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeHandRequest'
      responses:
        '200':
          description: V1 envelope; no deck or collection writes. Draw/export/analysis echo exact revision and inputKey.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/HandAnalysisResult'
                  meta:
                    type: object
                  errors:
                    type: array
                    items:
                      type: object
        '400':
          description: VALIDATION_ERROR or DRAFT_STRUCTURE_INVALID
        '401':
          description: UNAUTHORIZED
        '403':
          description: FORBIDDEN; ADMIN player required
        '429':
          description: RATE_LIMITED
        '503':
          description: DRAFT_PREVIEW_UNAVAILABLE; retain draft, explicit retry
...